+++ resources/views/page.blade.php
@@ -5,6 +5,10 @@
   @endphp
   @while(have_posts())
     {!! the_post() !!}
+    @php
+      global $post;
+      extract(\App\TemplateHelpers::getPostData($post->ID));
+    @endphp
     @include('patterns.02-organisms.content.content-page')
   @endwhile
 @endsection