+++ resources/views/patterns/02-organisms/global/header.blade.php
@@ -14,8 +14,13 @@
     $header_logo_class = "";
   }
 
-  if (carbon_get_the_post_meta('hero_type') == 'full_overlay') {
-    $header_class = " c-header--overlay u-theme--gradient--top";
+  $header_type = carbon_get_the_post_meta('head_type');
+  if (carbon_get_the_post_meta('hero_type') == 'full_overlay' || $header_type == 'overlay' || $header_type == 'overlay_gradient') {
+    if (carbon_get_the_post_meta('head_type') == 'overlay'){
+      $header_class = " c-header--overlay";
+    }else{
+      $header_class = " c-header--overlay u-theme--gradient--top";
+    }
     $logo_class = "u-path-fill--white";
   } else {
     $header_class = "";
@@ -27,6 +32,8 @@
       $logoContainerClass[] = 'c-header__logo--wide';
   }
 @endphp
+
+@if ($header_type != 'remove')
 <header class="c-header{{ $header_class }}" role="banner" id="header">
   <div class="c-header--inner">
     <div class="c-header__nav-secondary">
@@ -59,4 +66,5 @@
     </div>
   @endif
 </header> <!-- .c-header -->
+@endif
 @include('patterns.01-molecules.navigation.drawer-navigation')